+++ CHANGELOG.md
@@ -1,17 +1,23 @@
 CHANGELOG
 =========
 
+## v0.4.0 - *2/7/2018*
+
+**Note:** hardcoding the architecture to amd64.
+
+- Support for determining local operating systems *(Darwin, FreeBSD, Linux, NetBSD, OpenBSD, Solaris)*.
+
 ## v0.3.0 - *2/3/2018*
 
-**Note:** currently hardcoding the platform to darwin_amd64.
+**Note:** hardcoding the os/architecture to darwin_amd64.
 
 - Installation instructions set's `$HASHIPM_ROOT` environment variable. Hashipm uses this variable to resolve paths for packages and sourcing `lib/yaml.sh`.
 
 ## v0.2.0 - *1/31/2018*
 
-**Note:** currently hardcoding the platform to darwin_amd64.
+**Note:** hardcoding the os/architecture to darwin_amd64.
 
-- Initial darwin_amd64 platform support for Consul, Packer, Terraform, and Vault.
+- Initial Darwin amd64 support for Consul, Packer, Terraform, and Vault.
 
 ## v0.1.0 - *1/29/2018*